requirements.txt
asgiref==3.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==5.0.3
django-filter==24.1
djangorestframework==3.15.0
httpie==3.2.2
idna==3.7
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
mysqlclient==2.2.4
Pygments==2.17.2
PySocks==1.7.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
